+++ osgp-adapter-domain-smartmetering/src/main/java/com/alliander/osgp/adapter/domain/smartmetering/infra/jms/core/messageprocessors/PeriodicMeterReadsresponseMessageProcessor.java
@@ -19,7 +19,7 @@ import org.springframework.stereotype.Component;
 import com.alliander.osgp.adapter.domain.smartmetering.application.services.MonitoringService;
 import com.alliander.osgp.adapter.domain.smartmetering.infra.jms.core.OsgpCoreResponseMessageProcessor;
 import com.alliander.osgp.domain.core.valueobjects.DeviceFunction;
-import com.alliander.osgp.dto.valueobjects.smartmetering.PeriodicMeterReads;
+import com.alliander.osgp.dto.valueobjects.smartmetering.PeriodicMeterReadsContainer;
 import com.alliander.osgp.shared.exceptionhandling.OsgpException;
 import com.alliander.osgp.shared.infra.jms.Constants;
 import com.alliander.osgp.shared.infra.jms.ResponseMessage;
@@ -55,7 +55,7 @@ public class PeriodicMeterReadsresponseMessageProcessor extends OsgpCoreResponse
         ResponseMessage responseMessage = null;
         ResponseMessageResultType responseMessageResultType = null;
         OsgpException osgpException = null;
-        PeriodicMeterReads periodicMeterReads = null;
+        PeriodicMeterReadsContainer periodicMeterReadsContainer = null;
 
         try {
             correlationUid = message.getJMSCorrelationID();
@@ -66,7 +66,7 @@ public class PeriodicMeterReadsresponseMessageProcessor extends OsgpCoreResponse
             responseMessage = (ResponseMessage) message.getObject();
             responseMessageResultType = responseMessage.getResult();
             osgpException = responseMessage.getOsgpException();
-            periodicMeterReads = (PeriodicMeterReads) responseMessage.getDataObject();
+            periodicMeterReadsContainer = (PeriodicMeterReadsContainer) responseMessage.getDataObject();
         } catch (final JMSException e) {
             LOGGER.error("UNRECOVERABLE ERROR, unable to read ObjectMessage instance, giving up.", e);
             LOGGER.debug("correlationUid: {}", correlationUid);
@@ -83,7 +83,7 @@ public class PeriodicMeterReadsresponseMessageProcessor extends OsgpCoreResponse
             LOGGER.info("Calling application service function to handle response: {}", messageType);
 
             this.monitoringService.handlePeriodicMeterReadsresponse(deviceIdentification, organisationIdentification,
-                    correlationUid, messageType, responseMessageResultType, osgpException, periodicMeterReads);
+                    correlationUid, messageType, responseMessageResultType, osgpException, periodicMeterReadsContainer);
 
         } catch (final Exception e) {
             this.handleError(e, correlationUid, organisationIdentification, deviceIdentification, messageType);
